Dev & Train NLP Covid Data
In data/coronavirus, we have provided a second dataset consisting of reddit comments on posts related to the COVID-19 pandemic early last year. Each example is a single line in the CSV file, consisting of a comment and associated sentiment label: 0 for negative, and 1 for positive. This dataset consists of around 4 times as many examples as the triage/disaster dataset. The task on this dataset is to, given the text of a comment, predict its sentiment label.